architect/requirements (1).txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
ollama>=0.1.7
pydantic>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
langgraph>=0.1.0
langchain-core>=0.2.0
temporalio>=1.6.0
